+++ srcs/json/json_object.c
@@ -6,7 +6,7 @@
 /*   By: llelievr <llelievr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/16 23:34:20 by llelievr          #+#    #+#             */
-/*   Updated: 2019/03/01 15:35:22 by llelievr         ###   ########.fr       */
+/*   Updated: 2019/04/08 14:30:32 by llelievr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -93,7 +93,7 @@ t_json_value		*json_object_get(t_json_object *obj, char *key)
 	mem = obj->elements;
 	while (mem)
 	{
-		if (mem->string->value_len == ft_strlen(key)
+		if (mem->string->value_len == (int)ft_strlen(key)
 			&& ft_strncmp(mem->string->value, key, mem->string->value_len) == 0)
 			return (mem->value);
 		mem = mem->next;
